Use signal handler and monitor thread to prevent lock on SIGINT (#6)
To prevent the time series from blocking indefinitely on SIGINT (because
it is waiting for some condition to be fulfilled, e.g. the next element
to be added), add a monitor thread that releases the condition wait when
a SIGINT is detected.
Further add a check in all wait loops and throw an exception (as there
is usually no valid return value if SIGINT is received before the
condition is fulfilled).