Provide create_* methods for shared pointers
There are some issues with moving an instance into a pointer, so
`std::make_shared<..>(MultiprocessTimeSeries::create_leader(..));` is
currently not working.  Until this is resolved, simply add additional
`create_*_ptr` factories which directly return a shared_ptr.

Move the code to get max_length and start index for the follower to a
separate method to avoid too much duplication.